+++ src/controller/likeController.ts
@@ -1,11 +1,11 @@
 import { ArticleIdParams } from '../structs/articleStruct.js';
-import { ProductParams } from '../structs/productStruct.js';
+import { ProductIdParams } from '../structs/productStruct.js';
 import { HttpError } from '../utils/errors.js';
 import { prisma } from '../utils/prismaClient.js';
 import { Request, Response } from 'express';
 export class Like {
   static toggleProductLike = async (req: Request, res: Response) => {
-    const { productId } = ProductParams.create(req.params);
+    const { productId } = ProductIdParams.create(req.params);
     const user = req.user;
     if (!user) {
       throw new HttpError(401, '잘못된 접근입니다.');
