+++ src/middleware/authenticate.ts
@@ -8,8 +8,7 @@ export const authenticate = async (req: Request, res: Response, next: NextFuncti
     if (!accessToken) {
       return res.status(401).send({ message: '유효하지 않은 접근입니다.' });
     }
-    const { userId: userIdString } = verifyAccessToken(accessToken);
-    const userId = parseInt(userIdString, 10);
+    const { userId: userId } = verifyAccessToken(accessToken);
     const user = await prisma.user.findUnique({ where: { id: userId } });
     if (!user) {
       return res.status(401).send({ message: '아이디 또는 비밀번호가 일치하지 않습니다' });
