+++ src/routers/orderRouter.ts
@@ -1,0 +1,12 @@
+// import express from 'express';
+// import { validate } from '../middleware/validate.js';
+// import { CreateOrder } from '../structs/orderStruct.js';
+// import { tryCatchHandler } from '../middleware/errorhandler.js';
+// import { prisma } from '../utils/prismaClient.js';
+// import { OrderController } from '../controller/orderController.js';
+// const orderRouter = express.Router();
+
+// orderRouter.route('/').post(validate(CreateOrder), tryCatchHandler(OrderController.createOrder));
+
+// export default orderRouter;
+//시간 여유가 있으면 order 컨트롤러와 같이 한번에 수정 예정
